move ManagerType to core module
