+++ eyes.sdk.core/src/main/java/com/applitools/eyes/universal/ManagerType.java
@@ -1,4 +1,4 @@
-package com.applitools.eyes.selenium;
+package com.applitools.eyes.universal;
 
 /**
  * manager type
